Feat : Payment-infra 내부 DB 초기설정
