+++ backoffice-manage/build.gradle.kts
@@ -30,6 +30,7 @@ dependencies {
     // DB
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     runtimeOnly("com.h2database:h2") // 개발용 인메모리 DB
+    runtimeOnly("org.postgresql:postgresql:42.7.3")
 
     // 공용 모듈
     implementation(project(":common"))
