+++ common/build.gradle.kts
@@ -23,6 +23,9 @@ dependencies {
 	implementation("org.springframework.boot:spring-boot-starter-web")
 	implementation("org.springframework.boot:spring-boot-starter-aop")
 
+	// jakarta
+	implementation("jakarta.validation:jakarta.validation-api:3.0.2")
+
 	// ✅ Lombok
 	compileOnly("org.projectlombok:lombok")
 	annotationProcessor("org.projectlombok:lombok")
