+++ payment-infra/src/main/java/com/fastcampus/paymentinfra/entity/CardInfo.java
@@ -1,0 +1,31 @@
+package com.fastcampus.paymentinfra.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table
+@Data
+public class CardInfo {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long cardInfoId;
+
+    private Long userId;
+    private String type;
+    private String token;
+    private String last4;
+    private String cardCompany;
+
+    @CreationTimestamp
+    @Column(columnDefinition = "TIMESTAMP")
+    private LocalDateTime createdAt;
+
+    @UpdateTimestamp
+    @Column(columnDefinition = "TIMESTAMP")
+    private LocalDateTime updatedAt;
+}
