+++ payment-infra/src/main/java/com/fastcampus/paymentinfra/entity/Keys.java
@@ -1,0 +1,28 @@
+package com.fastcampus.paymentinfra.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table
+@Data
+public class Keys {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long keysId;
+
+    private String encryptedKey;
+    private Long merchantId;
+
+    @CreationTimestamp
+    @Column(columnDefinition = "TIMESTAMP")
+    private LocalDateTime createdAt;
+
+    @UpdateTimestamp
+    @Column(columnDefinition = "TIMESTAMP")
+    private LocalDateTime updatedAt;
+}
