+++ payment-infra/src/main/java/com/fastcampus/paymentinfra/entity/Payment.java
@@ -1,0 +1,39 @@
+package com.fastcampus.paymentinfra.entity;
+
+
+import jakarta.persistence.*;
+import lombok.Data;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table
+@Data
+public class Payment {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long paymentId;
+
+    private Long transactionId;
+    private Long userId;
+    private Long paymentMethod;
+    private String paymentStatus;
+    private Long paidAmount;
+
+    @CreationTimestamp
+    @Column(columnDefinition = "TIMESTAMP")
+    private LocalDateTime approvedAt;
+
+    private String failReason;
+    private Long lastTransactionId;
+
+    @CreationTimestamp
+    @Column(columnDefinition = "TIMESTAMP")
+    private LocalDateTime createdAt;
+
+    @UpdateTimestamp
+    @Column(columnDefinition = "TIMESTAMP")
+    private LocalDateTime updatedAt;
+}
