--- payment-infra/src/main/java/com/fastcampus/paymentinfra/infra/entity/CommonEntity.java
@@ -1,4 +1,0 @@
-package com.fastcampus.paymentinfra.infra.entity;
-
-public class CommonEntity {
-}
