--- payment-infra/src/main/java/com/fastcampus/paymentinfra/infra/entity/PaymentMethod.java
@@ -1,28 +1,0 @@
-package com.fastcampus.paymentinfra.infra.entity;
-
-
-import jakarta.persistence.*;
-import lombok.Data;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table
-@Data
-public class PaymentMethod {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long paymentMethodId;
-
-    private String type;
-
-    @CreationTimestamp
-    @Column(columnDefinition = "TIMESTAMP")
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    @Column(columnDefinition = "TIMESTAMP")
-    private LocalDateTime updatedAt;
-}
