--- payment-infra/src/main/java/com/fastcampus/paymentinfra/infra/entity/Users.java
@@ -1,30 +1,0 @@
-package com.fastcampus.paymentinfra.infra.entity;
-
-import jakarta.persistence.*;
-import lombok.Data;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table
-@Data
-public class Users {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long userId;
-
-    private String name;
-    private String phone;
-    private String email;
-    private String simplePassword;
-
-    @CreationTimestamp
-    @Column(columnDefinition = "TIMESTAMP")
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    @Column(columnDefinition = "TIMESTAMP")
-    private LocalDateTime updatedAt;
-}
