+++ payment-infra/src/main/resources/application-dev.yml
@@ -1,0 +1,14 @@
+spring:
+  application:
+    name: payment-infra
+
+  datasource:
+    url: jdbc:postgresql://localhost:5432/testdb
+    driver-class-name: org.postgresql.Driver
+    username: testuser
+    password: testpass
+
+  jpa:
+    database-platform: org.hibernate.dialect.PostgreSQLDialect
+    hibernate:
+      ddl-auto: update
