--- payment-infra/src/main/resources/application.properties
@@ -1,7 +1,0 @@
-spring.application.name=payment-infra
-
-spring.datasource.url= jdbc:h2:mem:testdb
-spring.datasource.driver-class-name= org.h2.Driver
-spring.datasource.username= sa
-spring.datasource.password= 1234
-jpa.database-platform= org.hibernate.dialect.H2Dialect
